downgrade github.com/matttproud/golang_protobuf_extensions to v1.0.1
This dependency was updated in 6b864dc4e10bf62508ecbfa3106ac7952004ea9e, but using
a version that's 1 commit ahead of the latest (v1.0.1) release. As a consequence,
any project that uses k8s.io/component-base as a dependency, but that prefers to
use tagged releases, will be required to add a `replace` rule in their `go.mod`.

Looking at the difference between the currently used version and the last release,
the only difference is the addition of a `go.mod`:
https://github.com/matttproud/golang_protobuf_extensions/compare/v1.0.1...c182affec369

While it can help to have a `go.mod` to resolve the minimum version of (indirect)
dependencies, in this case, the (minimum) versions of dependencies defined in this
module's `go.mod` are quite a bit behind the version defined in k8s' own go.mod,
so would already be up-to-date.

Based on the above, it looks like the `go.mod` (currently) doesn't provide much
value, and it would be safe to roll back 1 commit to use the tagged/released
version of this module.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

Kubernetes-commit: 82a765c52fb94401a1b85c9fb9d2b1e457e84556