fix openapi/v3 non local apiservices aggregation
When resyncing APIServices (every minute), the AggregationController calls
AddUpdateAPIService immediately, but it may take some time before the
reconciliation of the controller calls UpdateAPIServiceSpec. This results in a
temporary downtime of discovery of these APIServices, which will stop reporting
of these in the aggregated openapi/v3 by specProxier. This fix ensures that old
discovery is kept during a time, when the AggregationController is reconciling.

Kubernetes-commit: 7e13299aa995a1ceb609e44b9fb7cd2468390bca